+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 
 ### Fixed
 - Fixed hierarchical structure being lost when copying nodes to another site ([#452](https://github.com/verbb/navigation/issues/452)).
+- Fixed `node.children` being empty on front-end reads after copying nodes to another site in multisite menus ([#452](https://github.com/verbb/navigation/issues/452)).
+- Fixed `node.children` being empty on cached level-scoped node queries (for example `.level(1)`) after the first page load ([#452](https://github.com/verbb/navigation/issues/452)).
 - Fixed an install error around `ProjectConfig::onDelete`. (thanks @bramnijssen)
 - Fixed `craft.navigation.nodes()` string shorthand and deprecated node query criteria keys setting an unknown `menuHandle` property on `NodeQuery`. (thanks @bramnijssen)